add UI components for home page
