+++ src/app/servises/firestoreService.js
@@ -1,0 +1,9 @@
+import { db } from "../../firebase";
+
+// CREATE
+
+// READ
+
+// UPDATE
+
+// DELETE
