--- src/app/servises/post-api.js
@@ -1,29 +1,0 @@
-// import { db } from "../../firebase";
-// import { collection, addDoc, doc, setDoc } from "firebase/firestore";
-
-// const COLLECTION_NAME = "posts";
-// const postCollectionRef = collection();
-
-// // CREATE
-// // export const addPost = async (data) => {
-// //   try {
-// //     await addDoc(colRef, data);
-// //     alert("Success!");
-// //   } catch (error) {
-// //     alert(error);
-// //   }
-// // };
-
-// // READ
-
-// // UPDATE
-// // export const editPost = async (data) => {
-// //   try {
-// //     const postRef = doc(db, "posts");
-// //     await setDoc(newCityRef, data);
-// //   } catch (error) {
-// //     alert(error);
-// //   }
-// // };
-
-// // DELETE
