+++ src/features/Home/Home.js
@@ -1,0 +1,13 @@
+import React from "react";
+
+import { Hero } from "../../components/organisms/Hero";
+import { HomeActions } from "../../components/organisms/HomeActions";
+
+export const Home = () => {
+  return (
+    <>
+      <Hero />
+      <HomeActions />
+    </>
+  );
+};
